Move `TenureChangeCause` and `TenureChangePayload`
